feat: Convert Events into extension functions
