+++ src/main/kotlin/Novu.kt
@@ -14,12 +14,7 @@ import co.novu.api.NotificationTemplatesApi
 import co.novu.api.NotificationsApi
 import co.novu.api.SubscribersApi
 import co.novu.api.TopicsApi
-import co.novu.dto.request.BroadcastEventRequest
-import co.novu.dto.request.BulkTriggerEventRequest
-import co.novu.dto.request.TriggerEventRequest
 import co.novu.helpers.RetrofitHelper
-import kotlinx.coroutines.runBlocking
-import mu.KotlinLogging
 
 data class NovuConfig(
     var backendUrl: String = "https://api.novu.co/v1/",
@@ -32,11 +27,9 @@ class Novu(
 
     constructor(apiKey: String) : this(NovuConfig(apiKey = apiKey))
 
-    private val logger = KotlinLogging.logger {}
-
     private val retrofitInstance = RetrofitHelper(config).getInstance()
 
-    private val eventsApi = retrofitInstance.create(EventsApi::class.java)
+    internal val eventsApi = retrofitInstance.create(EventsApi::class.java)
 
     internal val subscribersApi = retrofitInstance.create(SubscribersApi::class.java)
 
@@ -63,39 +56,4 @@ class Novu(
     internal val notificationGroupsApi = retrofitInstance.create(NotificationGroupsApi::class.java)
 
     internal val inboundParseApi = retrofitInstance.create(InboundParseApi::class.java)
-    fun trigger(body: TriggerEventRequest) = runBlocking {
-        val response = eventsApi.triggerEvent(body)
-        if (response.isSuccessful) {
-            response.body().apply { logger.debug { this } }
-        } else {
-            throw Exception(response.errorBody()?.string())
-        }
-    }
-
-    fun bulkTrigger(body: BulkTriggerEventRequest) = runBlocking {
-        val response = eventsApi.bulkTriggerEvent(body)
-        if (response.isSuccessful) {
-            response.body().apply { logger.debug { this } }
-        } else {
-            throw Exception(response.errorBody()?.string())
-        }
-    }
-
-    fun broadcast(body: BroadcastEventRequest) = runBlocking {
-        val response = eventsApi.broadcastEvent(body)
-        if (response.isSuccessful) {
-            response.body().apply { logger.debug { this } }
-        } else {
-            throw Exception(response.errorBody()?.string())
-        }
-    }
-
-    fun cancelTriggerEvent(transactionId: String) = runBlocking {
-        val response = eventsApi.cancelTriggerEvent(transactionId)
-        if (response.isSuccessful) {
-            response.body().apply { logger.debug { this } }
-        } else {
-            throw Exception(response.errorBody()?.string())
-        }
-    }
 }
